(fix) SJT-132 Create the Edit Payment mode modal
