+++ packages/esm-billing-app/src/index.ts
@@ -27,7 +27,6 @@ import { ClockIn } from './payment-points/payment-point/clock-in.component';
 import { ClockOut } from './payment-points/payment-point/clock-out.component';
 import DeletePaymentModeModal from './payment-modes/delete-payment-mode.modal';
 import PaymentModeWorkspace from "./payment-modes/payment-mode.workspace";
-import EditPaymentMode from './payment-modes/edit-payment-mode.modal';
 
 const moduleName = '@ehospital/esm-billing-app';
 
@@ -84,7 +83,6 @@ export const paymentModesPanelLink = getSyncLifecycle(
 
 export const paymentModeWorkspace = getSyncLifecycle(PaymentModeWorkspace, options);
 export const deletePaymentModeModal = getSyncLifecycle(DeletePaymentModeModal, options);
-export const editPaymentMode = getSyncLifecycle(EditPaymentMode, options)
 
 export const importTranslation = require.context('../translations', false, /.json$/, 'lazy');
