+++ packages/esm-billing-app/src/payment-modes/payment-mode-dashboard.component.tsx
@@ -55,16 +55,11 @@ const PaymentModeDashboard: React.FC<PaymentModeDashboardProps> = () => {
     });
   };
 
-  const showEditPaymentModeModal = (paymentMode: PaymentMode) => {
-    const dispose = showModal('edit-payment-mode', {
-      closeModal: () => dispose(),
-      initialPaymentMode: paymentMode,
-    });
-  };
-
-  const createPaymentModeModal = () => {
+  const createPaymentModeModal = (paymentMode: PaymentMode, modalTitle: string) => {
     const dispose = showModal('create-payment-mode', {
           closeModal: () => dispose(),
+          initialPaymentMode: paymentMode,
+          modalTitle,
         });
   }
 
@@ -119,10 +114,9 @@ const PaymentModeDashboard: React.FC<PaymentModeDashboardProps> = () => {
     <div>
       <CardHeader title="Payment Modes">
         <Button
-          onClick={() => {createPaymentModeModal()}}
-          className={styles.addPaymentModeButton}
-          size="md"
-          kind="ghost">
+          onClick={() => {createPaymentModeModal(paymentModes[''], t('createPaymentMode', 'Create Payment Mode'))}}
+          className={styles.createPaymentModeButton}
+          size="md">
           {t('addPaymentMode', 'Add Payment Mode')}
         </Button>
       </CardHeader>
@@ -167,7 +161,7 @@ const PaymentModeDashboard: React.FC<PaymentModeDashboardProps> = () => {
                         <TableCell className="cds--table-column-menu">
                           <OverflowMenu size={size} iconDescription={t('actions', 'Actions')} flipped>
                             <OverflowMenuItem
-                              onClick={() => showEditPaymentModeModal(paymentModes[index]) }
+                              onClick={() => createPaymentModeModal(paymentModes[index], t('editPaymentMode', 'Edit Payment Mode')) }
                               itemText={t('edit', 'Edit')}
                             />
                             <OverflowMenuItem
